+++ sofa-boot-project/sofa-boot-starters/runtime-sofa-boot-starter/src/test/java/com/alipay/sofa/runtime/test/ambush/JvmFilterInterruptedConfiguration.java
@@ -16,8 +16,8 @@
  */
 package com.alipay.sofa.runtime.test.ambush;
 
-import com.alipay.sofa.runtime.ambush.Context;
-import com.alipay.sofa.runtime.ambush.JvmFilter;
+import com.alipay.sofa.runtime.filter.JvmFilterContext;
+import com.alipay.sofa.runtime.filter.JvmFilter;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ImportResource;
@@ -36,7 +36,7 @@ public class JvmFilterInterruptedConfiguration {
     public JvmFilter egressFilter1() {
         return new JvmFilter() {
             @Override
-            public boolean before(Context context) {
+            public boolean before(JvmFilterContext context) {
                 ++beforeCount;
                 return true;
             }
@@ -47,7 +47,7 @@ public class JvmFilterInterruptedConfiguration {
             }
 
             @Override
-            public boolean after(Context context) {
+            public boolean after(JvmFilterContext context) {
                 return true;
             }
         };
@@ -57,7 +57,7 @@ public class JvmFilterInterruptedConfiguration {
     public JvmFilter egressFilter2() {
         return new JvmFilter() {
             @Override
-            public boolean before(Context context) {
+            public boolean before(JvmFilterContext context) {
                 ++beforeCount;
                 return false;
             }
@@ -68,7 +68,7 @@ public class JvmFilterInterruptedConfiguration {
             }
 
             @Override
-            public boolean after(Context context) {
+            public boolean after(JvmFilterContext context) {
                 return true;
             }
         };
@@ -78,7 +78,7 @@ public class JvmFilterInterruptedConfiguration {
     public JvmFilter egressFilter3() {
         return new JvmFilter() {
             @Override
-            public boolean before(Context context) {
+            public boolean before(JvmFilterContext context) {
                 ++beforeCount;
                 context.setInvokeResult("interrupted");
                 return false;
@@ -90,7 +90,7 @@ public class JvmFilterInterruptedConfiguration {
             }
 
             @Override
-            public boolean after(Context context) {
+            public boolean after(JvmFilterContext context) {
                 return true;
             }
         };
