+++ sofa-boot-project/sofa-boot-starters/runtime-sofa-boot-starter/src/test/java/com/alipay/sofa/runtime/test/ambush/JvmFilterTest.java
@@ -16,9 +16,9 @@
  */
 package com.alipay.sofa.runtime.test.ambush;
 
-import com.alipay.sofa.runtime.ambush.Context;
-import com.alipay.sofa.runtime.ambush.FilterHolder;
-import com.alipay.sofa.runtime.ambush.JvmFilter;
+import com.alipay.sofa.runtime.filter.JvmFilterContext;
+import com.alipay.sofa.runtime.filter.JvmFilterHolder;
+import com.alipay.sofa.runtime.filter.JvmFilter;
 import com.alipay.sofa.runtime.test.RuntimeTestBase;
 import org.junit.Assert;
 import org.junit.BeforeClass;
@@ -39,29 +39,29 @@ public class JvmFilterTest extends RuntimeTestBase {
 
     @BeforeClass
     public static void before() {
-        FilterHolder.clearJvmFilters();
+        JvmFilterHolder.clearJvmFilters();
     }
 
     @Test
     public void test() {
         Assert.assertEquals("egressFilter2", myService.say());
-        Assert.assertEquals(5, FilterHolder.getJvmFilters().size());
+        Assert.assertEquals(5, JvmFilterHolder.getJvmFilters().size());
         Assert.assertEquals(3, JvmFilterTestConfiguration.beforeCount);
         Assert.assertEquals(1, JvmFilterTestConfiguration.afterCount);
     }
 
     @Test
     public void testResort() {
-        FilterHolder.clearJvmFilters();
+        JvmFilterHolder.clearJvmFilters();
 
-        FilterHolder.addJvmFilter(new JvmFilter() {
+        JvmFilterHolder.addJvmFilter(new JvmFilter() {
             @Override
-            public boolean before(Context context) {
+            public boolean before(JvmFilterContext context) {
                 return false;
             }
 
             @Override
-            public boolean after(Context context) {
+            public boolean after(JvmFilterContext context) {
                 return false;
             }
 
@@ -71,14 +71,14 @@ public class JvmFilterTest extends RuntimeTestBase {
             }
         });
 
-        FilterHolder.addJvmFilter(new JvmFilter() {
+        JvmFilterHolder.addJvmFilter(new JvmFilter() {
             @Override
-            public boolean before(Context context) {
+            public boolean before(JvmFilterContext context) {
                 return false;
             }
 
             @Override
-            public boolean after(Context context) {
+            public boolean after(JvmFilterContext context) {
                 return false;
             }
 
@@ -88,18 +88,18 @@ public class JvmFilterTest extends RuntimeTestBase {
             }
         });
 
-        List<JvmFilter> filters = FilterHolder.getJvmFilters();
+        List<JvmFilter> filters = JvmFilterHolder.getJvmFilters();
         Assert.assertEquals(9, filters.get(0).getOrder());
         Assert.assertEquals(10, filters.get(1).getOrder());
 
-        FilterHolder.addJvmFilter(new JvmFilter() {
+        JvmFilterHolder.addJvmFilter(new JvmFilter() {
             @Override
-            public boolean before(Context context) {
+            public boolean before(JvmFilterContext context) {
                 return false;
             }
 
             @Override
-            public boolean after(Context context) {
+            public boolean after(JvmFilterContext context) {
                 return false;
             }
 
@@ -109,7 +109,7 @@ public class JvmFilterTest extends RuntimeTestBase {
             }
         });
 
-        filters = FilterHolder.getJvmFilters();
+        filters = JvmFilterHolder.getJvmFilters();
         Assert.assertEquals(0, filters.get(0).getOrder());
     }
 }
